+++ CHANGELOG.md
@@ -145,6 +145,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * The new `jj sign` command allows signing commits.
 
+* The new `jj unsign` command allows unsigning commits.
+
 ### Fixed bugs
 
 * `jj git fetch` with multiple remotes will now fetch from all remotes before